# openapi.yaml
openapi: 3.0.3
info:
  title: Transparency Dashboard API
  description: API for tracking grants and spending on-chain for organizational transparency
  version: 1.0.0
  contact:
    name: Ethereum Community Foundation
servers:
  - url: /api/v1
    description: API server
paths:
...
  /api/v1/expenses:
    get:
      summary: Get expenses
      description: Returns a list of organizational expenses
      operationId: getExpenses
      tags:
        - Expenses
      parameters:
        - name: category
          in: query
          description: Filter expenses by category
          required: false
          schema:
            type: string
            example: "travel"
        - name: limit
          in: query
          description: Maximum number of expenses to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of expenses to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Successfully retrieved expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create expense
      description: Creates a new organizational expense
      operationId: createExpense
      tags:
        - Expenses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                item:
                  type: string
                  description: Description of the expense item
                  example: "Flight to conference"
                quantity:
                  type: integer
                  description: Quantity of items
                  example: 1
                price:
                  type: string
                  description: Price per item
                  example: "450.00"
                purpose:
                  type: string
                  description: Purpose of the expense
                  example: "Attending Ethereum conference for research"
                category:
                  type: string
                  description: Expense category
                  example: "travel"
                date:
                  type: string
                  format: date
                  description: Date of the expense
                  example: "2023-12-15"
                txHash:
                  type: string
                  description: Optional transaction hash if paid on-chain
                  example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
                  nullable: true
              required:
                - item
                - quantity
                - price
                - purpose
                - category
                - date
      responses:
        '201':
          description: Expense created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/expenses/{id}:
    put:
      summary: Update expense
      description: Updates an existing organizational expense
      operationId: updateExpense
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The expense ID
          schema:
            type: string
            example: "exp_12345"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                item:
                  type: string
                  description: Description of the expense item
                  example: "Flight to conference"
                quantity:
                  type: integer
                  description: Quantity of items
                  example: 1
                price:
                  type: number
                  description: Price per item
                  example: 450.00
                purpose:
                  type: string
                  description: Purpose of the expense
                  example: "Attending Ethereum conference for research"
                category:
                  type: string
                  description: Expense category
                  example: "travel"
                date:
                  type: string
                  format: date
                  description: Date of the expense
                  example: "2023-12-15"
                txHash:
                  type: string
                  description: Optional transaction hash if paid on-chain
                  example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
                  nullable: true
              required:
                - item
                - quantity
                - price
                - purpose
                - category
                - date
      responses:
        '200':
          description: Expense updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete expense
      description: Deletes an organizational expense
      operationId: deleteExpense
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The expense ID
          schema:
            type: string
            example: "exp_12345"
      responses:
        '204':
          description: Expense deleted successfully
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/expenses/{expenseId}/receipts/{receiptId}:
    delete:
      summary: Delete receipt
      description: Deletes a receipt document from an expense
      operationId: deleteReceipt
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: expenseId
          in: path
          required: true
          description: The expense ID
          schema:
            type: string
            example: "exp_12345"
        - name: receiptId
          in: path
          required: true
          description: The receipt UUID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '204':
          description: Receipt deleted successfully
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Expense or receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Failed to fetch treasury data"
        code:
          type: string
          description: Error code
          example: "TREASURY_FETCH_ERROR"
      required:
        - error
        - code
...
    Expense:
      type: object
      properties:
        id:
          type: string
          description: Unique expense identifier
          example: "exp_12345"
        item:
          type: string
          description: Description of the expense item
          example: "Flight to conference"
        quantity:
          type: integer
          description: Quantity of items
          example: 1
        price:
          type: number
          description: Price per item
          example: 450.00
        purpose:
          type: string
          description: Purpose of the expense
          example: "Attending Ethereum conference for research"
        category:
          type: string
          description: Expense category
          example: "travel"
        receipts:
          type: array
          description: List of receipt documents
          items:
            $ref: '#/components/schemas/Receipt'
        date:
          type: string
          format: date
          description: Date of the expense
          example: "2023-12-15"
        txHash:
          type: string
          description: Optional transaction hash if paid on-chain
          example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
          nullable: true
      required:
        - id
        - item
        - quantity
        - price
        - purpose
        - category
        - receipts
        - date
    Receipt:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          description: Unique identifier for the receipt document
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: Display name of the receipt document
          example: "Flight Receipt - Conference 2023"
        downloadUrl:
          type: string
          format: uri
          description: URL to download the receipt document
          example: "https://example.com/receipts/download/550e8400-e29b-41d4-a716-446655440000"
      required:
        - uuid
        - name
        - downloadUrl
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...